Fixed Thumb URDF partially
J0 to J2 must be 55.4 mm in the hand manual p.19, but it was 57.6 mm in
the provided URDF file. J1-2 is 16.6.
I updated J0-1 and J1-2 values to sum upto 55.4 mm.
It is still not perfect: when the real thumb touches the palm,
visualized thumb does not touch it!